refactor: break up records.cli
