+++ requirements.txt
@@ -37,8 +37,8 @@ elasticsearch-dsl==2.2.0
 enum34==1.1.6
 Fabric==1.12.0
 fixture==1.5.8
-Flask==0.12.2
-Flask-Admin==1.5.1
+Flask==1.0.2
+Flask-Admin==1.5.2
 Flask-Alembic==2.0.1
 Flask-Assets==0.12
 Flask-Babel==0.11.1
@@ -140,7 +140,7 @@ marshmallow==2.13.5
 maxminddb==1.3.0
 maxminddb-geolite2==2018.308
 mock==2.0.0
-msgpack-python==0.4.8
+msgpack-python==0.5.6
 node-semver==0.1.1
 oauthlib==2.1.0
 ordereddict==1.1
