+++ scoap3/modules/records/ext.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import, print_function
 
-from scoap3.modules.compliance.cli import compliance
-from .cli import loadrecords, fixdb
+
+from .cli import loadrecords
 
 
 class Scoap3Records(object):
@@ -16,6 +16,4 @@ class Scoap3Records(object):
         """Flask application initialization."""
         app.extensions['scoap3-records'] = self
         app.cli.add_command(loadrecords)
-        app.cli.add_command(fixdb)
-        app.cli.add_command(compliance)
         return self
